+++ ItemActions/BetterPlayer.cs
@@ -18,6 +18,9 @@ using SpikysLib;
 using Terraria.GameContent.ItemDropRules;
 using System.Linq;
 using System.Text.RegularExpressions;
+using Microsoft.Xna.Framework;
+using System;
+using Terraria.GameContent;
 
 namespace BetterInventory.ItemActions;
 
@@ -39,6 +42,8 @@ public sealed class BetterPlayer : ModPlayer {
         On_ItemSlot.DyeSwap += HookDyeSwapFavorited;
         On_ItemSlot.ArmorSwap += HookArmorSwapFavorited;
         On_ItemSlot.EquipSwap += HookEquipSwapFavorited;
+
+        MonoModHooks.Add(typeof(ItemLoader).GetMethod(nameof(ItemLoader.ModifyTooltips)), HookHideTooltip);
     }
 
     public override void Unload() {
@@ -311,6 +316,60 @@ public sealed class BetterPlayer : ModPlayer {
         return str;
     }
 
+    private static List<TooltipLine> HookHideTooltip(Reflection.ItemLoader.ModifyTooltipsFn orig, Item item, ref int numTooltips, string[] names, ref string[] text, ref bool[] modifier, ref bool[] badModifier, ref int oneDropLogo, out Color?[] overrideColor, int prefixlineIndex) {
+        var tooltipsAll = orig.Invoke(item, ref numTooltips, names, ref text, ref modifier, ref badModifier, ref oneDropLogo, out overrideColor, prefixlineIndex);
+        if (!Configs.TooltipScroll.Enabled) return tooltipsAll;
+        if (_tooltipType != item.type) {
+            _tooltipType = item.type;
+            tooltipStart = 1;
+        }
+
+        int inset = Main.SettingsEnabled_OpaqueBoxBehindTooltips ? 18 : 4;
+        int count = Math.Max(1, (int)((Main.screenHeight-inset) * Configs.TooltipScroll.Value.maximumHeight / FontAssets.MouseText.Value.LineSpacing) - 2);
+        if (numTooltips <= count+1) return tooltipsAll;
+        HandleTooltipScroll(numTooltips, count);
+
+        (T[] text, T[]) NewArray<T>(T[] array) {
+            T[] a = new T[count+2];
+            a[0] = array[0];
+            return (array, a);
+        }
+        (var numTooltipsAll, numTooltips) = (numTooltips, count + 2);
+        List<TooltipLine> tooltips = new(count+2) { tooltipsAll[0] };
+        (var textAll, text) = NewArray(text);
+        (var modifierAll, modifier) = NewArray(modifier);
+        (var badModifierAll, badModifier) = NewArray(badModifier);
+        (var oneDropLogoAll, oneDropLogo) = (oneDropLogo, -1);
+        (var overrideColorAll, overrideColor) = NewArray(overrideColor);
+
+        for (int i = 0; i < count; i++) {
+            int iAll = tooltipStart + i;
+            tooltips.Add(tooltipsAll[iAll]);
+            text[i+1] = textAll[iAll];
+            modifier[i+1] = modifierAll[iAll];
+            badModifier[i+1] = badModifierAll[iAll];
+            if (oneDropLogoAll == iAll) oneDropLogo = i+1;
+            overrideColor[i+1] = overrideColorAll[iAll];
+        }
+
+        TooltipLine scrollLine = new(BetterInventory.Instance, "scrollTooltip", Language.GetTextValue($"{Localization.Keys.UI}.ScrollTooltip", tooltipStart, tooltipStart+count-1, numTooltipsAll-1)) { OverrideColor = Colors.RarityTrash };
+        tooltips.Add(scrollLine);
+        text[count+1] = scrollLine.Text;
+        overrideColor[count+1] = scrollLine.OverrideColor;
+
+        return tooltips;
+    }
+
+    private static void HandleTooltipScroll(int numTooltips, int visibleLines) {
+        PlayerInput.LockVanillaMouseScroll("BetterInventory: Scrollable Tooltip");
+        int scroll = PlayerInput.ScrollWheelDelta / 120;
+        if (scroll == 0) return;
+        int s = tooltipStart;
+        tooltipStart -= scroll;
+        tooltipStart = Math.Clamp(tooltipStart, 1, numTooltips - visibleLines);
+        if (s != tooltipStart) SoundEngine.PlaySound(SoundID.MenuTick);
+    }
+
     public const string VisibilityTag = "visibility";
     public const string RecipesTag = "recipes";
     public const string FavoritedInBanksTag = "favorited";
@@ -320,6 +379,9 @@ public sealed class BetterPlayer : ModPlayer {
     private readonly static List<TooltipLine> _bagContentTooltips = [];
 
     internal Item? _tempGuideTile;
+
+    private static int _tooltipType;
+    public static int tooltipStart;
 
     private static readonly Regex _coinRegex = new("""^(\[c\/[0-9a-fA-F]{6}:)?(\d+)( [a-zA-Z ]+)(\])?$""");
 }